Refactor backend app using express.router
